Add comments to analysis page
